nfd-topology-updater: manifests: fix /sys mount
Due to how ghw works, we need to mount the host sysfs
in a chroot friendly way. Nothing really changes besides
the mount paths inside the containers.

We need to repurpose the `--sysfs` option to be the chroot
 -not the full path on which sysfs is mounted.

Before:
--sysfs=/host-sysfs  -> sysfs root = /host-sysfs/

After:
--sysfs=/host        -> sysfs root = /host/sys/

(e.g. the "sys" part is implicitely and automatically added now - by GHW).

Unfortunately this is an API breakage, but since we didn't make any
backward compatibility promise yet, this is a breakage we can still
afford.

Signed-off-by: Francesco Romani <fromani@redhat.com>